refactor: change field entity submissions int to long

+++ src/main/java/com/su/ac/th/project/grader/entity/SubmissionsEntity.java
@@ -12,7 +12,7 @@ public class SubmissionsEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, nullable = false)
-    private int submissionId;
+    private Long submissionId;
 
     @Column(nullable = false)
     private Long userId;
